moves exists check to TankRouter fixes check existence bug in TankRouter

+++ src/routers/tankRouter.ts
@@ -94,20 +94,28 @@ TankRouter.put('/', async (req, res) => {
 
 TankRouter.delete('/:tank_id', async (req, res) => {
   const tank_id = req.params['tank_id'];
-  const response = await tankService.deleteTank(tank_id);
 
-  switch (response.message) {
-    case RESPONSE_MESSAGE.NO_ERROR:
-      res.status(204).send();
-      break;
-    case RESPONSE_MESSAGE.NOT_FOUND:
-      res.status(404).send(RESPONSE_MESSAGE.NOT_FOUND);
-      break;
-    case RESPONSE_MESSAGE.INTERNAL:
-      res.status(500).send(RESPONSE_MESSAGE.INTERNAL);
-      break;
-    default:
-      res.status(500).send(RESPONSE_MESSAGE.INTERNAL);
-      break;
+  const exists = await tankService.checkTankExists(tank_id);
+  if (!exists) {
+    res.status(404).send(RESPONSE_MESSAGE.NOT_FOUND);
+  }
+
+  if (exists) {
+    const response = await tankService.deleteTank(tank_id);
+
+    switch (response.message) {
+      case RESPONSE_MESSAGE.NO_ERROR:
+        res.status(204).send();
+        break;
+      case RESPONSE_MESSAGE.NOT_FOUND:
+        res.status(404).send(RESPONSE_MESSAGE.NOT_FOUND);
+        break;
+      case RESPONSE_MESSAGE.INTERNAL:
+        res.status(500).send(RESPONSE_MESSAGE.INTERNAL);
+        break;
+      default:
+        res.status(500).send(RESPONSE_MESSAGE.INTERNAL);
+        break;
+    }
   }
 });
